Extract run external command
